Created duplicates to fill in for cassie without stand
