+++ examples/cassie/controller/osc.cc
@@ -1,0 +1,1 @@
+#include "osc.h"
