+++ examples/cassie/controller/osc.h
@@ -1,0 +1,4 @@
+#ifndef OSC_HPP
+#define OSC_HPP
+
+#endif /* OSC_HPP */
